refactory utils.go and break it down into smaller files
