+++ testutil/simulation/rand_authority.go
@@ -1,0 +1,32 @@
+package simulation
+
+import (
+	"fmt"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
+
+	observertypes "github.com/zeta-chain/node/x/observer/types"
+)
+
+// GetPolicyAccount returns the simulation account associated with the policy address
+func GetPolicyAccount(ctx sdk.Context, k AuthorityKeeper, accounts []simtypes.Account) (simtypes.Account, error) {
+	policies, found := k.GetPolicies(ctx)
+	if !found {
+		return simtypes.Account{}, fmt.Errorf("policies object not found")
+	}
+	if len(policies.Items) == 0 {
+		return simtypes.Account{}, fmt.Errorf("no policies found")
+	}
+
+	admin := policies.Items[0].Address
+	address, err := observertypes.GetOperatorAddressFromAccAddress(admin)
+	if err != nil {
+		return simtypes.Account{}, err
+	}
+	simAccount, found := simtypes.FindAccount(accounts, address)
+	if !found {
+		return simtypes.Account{}, fmt.Errorf("admin account not found in list of simulation accounts")
+	}
+	return simAccount, nil
+}
